fix(ecommerce-stripe): package-name import, relocate import schemas, fix decl portability
- product.entity.ts: import sqlBaseProperties from @forklaunch/blueprint-core
  instead of a relative path into core (Rohin: 'suspicious' / 'cross package
  import'). blueprint-core is already a declared dependency.

- catalogImport.controller.ts: move the four Import*Schema definitions into
  domain/schemas/catalogImport.schema.ts (Rohin: 'should these live somewhere
  else?'). billing and iam controllers define zero schemas inline; this brings
  ecommerce in line.

- ecommerceMappers.types.ts: re-export the Base*Service mapper types so the
  emitted declaration can name them. ecommerce-stripe resolves its services
  from implementation-ecommerce-base (billing-stripe does not), so the
  container's inferred type reached into that package's internal paths and
  tsc raised TS2883 under the TS7 toolchain. These errors were previously
  masked by the rootDir failure that the rbac import fix cleared.

ecommerce-stripe now typechecks with 0 errors.
